Rename the importer to be less PopIt specific
This is more generally useful than just for importing from PopIt so
rename the module and class (to popolo_json and PopoloJSONImporter).
The old import and class name will still work.

This also updates the importer class docstring, e.g. to mention
EveryPolitician.
